refactor: Column uuid test move from e2e to integration.
